Rename syncmount to mount
